Moving API properties to separate file
